Update sha256example.html
Hey! Came across your project and just wanted to say thanks and let you know how valuable the work you are doing is!

Quick note: Today I was reading the SHA256 example on your website and I believe I may have come across a little error in the explanation.

"Peggy wants to prove that she knows a preimage for a digest chosen by Victor, without revealing what the preimage is"

My first intuition after reading this, given this phrasing, is that such a statement is not even really feasible. The setting as is basically forces Alice to perform a (second) preimage attack and find a preimage, or assumes that Victor knows a list of digests controlled by Alice. The first one is basically impossible, and the second is not mentioned in the text.

I know that explaining this type of thing is always extremely hard and I may be overcomplicating my take of this writeup, but just wanted to leave a comment regardless.

PS: This pull request basically contains what I would consider to be a more cryptographically correct phrasing, but do not feel like this is what an ideal correction would look like.